requirements.txt
deephaven-server
scipy
numba
numpy